api/requirements.txt
fastapi
uvicorn
python-dotenv
azure-cosmos